requirements.txt
fastapi>=0.115.12
uvicorn==0.27.1
python-dotenv>=1.1.0
websockets>=15.0.1
pydantic>=2.7.2,<3.0.0
pydantic-settings>=2.5.2
anthropic>=0.18.1
httpx>=0.27.0
starlette>=0.40.0,<0.47.0
anyio>=4.5.0
mcp>=1.8.0
aiohttp==3.9.3
fastmcp>=2.0.0 